requirements/base.txt
Django==4.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-extensions==3.2.3
celery==5.3.6
django-celery-results==2.5.1
redis==5.0.1
